Module fil_actors_shared::v10::runtime::policy::policy_constants

Constants

ADDRESSED_PARTITIONS_MAX

The maximum number of partitions that may be required to be loaded in a single invocation. This limits the number of simultaneous fault, recovery, or sector-extension declarations. With 48 deadlines (half-hour), 200 partitions per declaration permits loading a full EiB of 32 GiB sectors with 1 message per epoch within a single half-hour deadline. A miner can of course submit more messages.

ADDRESSED_SECTORS_MAX

The maximum number of sector infos that may be required to be loaded in a single invocation.

CALIBNET_MINIMUM_CONSENSUS_POWER

CHAIN_FINALITY

Epochs after which chain state is final with overwhelming probability (hence the likelihood of two fork of this size is negligible) This is a conservative value that is chosen via simulations of all known attacks.

CONSENSUS_FAULT_INELIGIBILITY_DURATION

Number of epochs after a consensus fault for which a miner is ineligible for permissioned actor methods and winning block elections.

DEAL_LIMIT_DENOMINATOR

Ratio of sector size to maximum deals per sector. The maximum number of deals is the sector size divided by this number (2^27) which limits 32 GiB sectors to 256 deals and 64 GiB sectors to 512

DEAL_UPDATES_INTERVAL

DealUpdatesInterval is the number of blocks between payouts for deals

DECLARATIONS_MAX

Maximum number of unique “declarations” in batch operations.

END_OF_LIFE_CLAIM_DROP_PERIOD

EXPIRED_PRE_COMMIT_CLEAN_UP_DELAY

The delay between pre-commit expiration and clean up from state. This enforces that expired pre-commits stay in state for a period of time creating a grace period during which a late-running aggregated prove-commit can still prove its non-expired pre-commits without resubmitting a message

FAULT_DECLARATION_CUTOFF

Minimum period before a deadline’s challenge window opens that a fault must be declared for that deadline. This lookback must not be less than WPoStChallengeLookback lest a malicious miner be able to selectively declare faults after learning the challenge value.

FAULT_MAX_AGE

The maximum age of a fault before the sector is terminated.

MAINNET_MINIMUM_CONSENSUS_POWER

MARKET_DEFAULT_ALLOCATION_TERM_BUFFER

MAXIMUM_VERIFIED_ALLOCATION_EXPIRATION

MAXIMUM_VERIFIED_ALLOCATION_TERM

MAX_AGGREGATED_PROOF_SIZE

Maximum total aggregated proof size.

MAX_AGGREGATED_SECTORS

Maximum amount of sectors that can be aggregated.

MAX_CONTROL_ADDRESSES

Maximum number of control addresses a miner may register.

MAX_MULTIADDR_DATA

MaxMultiaddrData is the maximum amount of data that can be stored in multiaddrs.

MAX_PARTITIONS_PER_DEADLINE

Maximum number of partitions that will be assigned to a deadline. For a minimum storage of up-to 1 EiB, we need 300 partitions per deadline. 48 * 32GiB * 2349 * 300 = 1.00808144 EiB So, to support up-to 10 Eib storage, we set this to 3000.

MAX_PEER_ID_LENGTH

MaxPeerIDLength is the maximum length allowed for any on-chain peer ID. Most Peer IDs are expected to be less than 50 bytes.

MAX_PRE_COMMIT_RANDOMNESS_LOOKBACK

MAX_REPLICA_UPDATE_PROOF_SIZE

Maximum total aggregated proof size.

MAX_SECTOR_EXPIRATION_EXTENSION

Maximum number of epochs past the current epoch a sector may be set to expire. The actual maximum extension will be the minimum of CurrEpoch + MaximumSectorExpirationExtension and sector.ActivationEpoch+sealProof.SectorMaximumLifetime()

MINIMUM_VERIFIED_ALLOCATION_SIZE

MINIMUM_VERIFIED_ALLOCATION_TERM

MIN_AGGREGATED_SECTORS

Minimum amount of sectors that can be aggregated.

MIN_SECTOR_EXPIRATION

Minimum number of epochs past the current epoch a sector may be set to expire.

NEW_SECTORS_PER_PERIOD_MAX

The maximum number of new sectors that may be staged by a miner during a single proving period.

PRE_COMMIT_CHALLENGE_DELAY

Number of epochs between publishing the pre-commit and when the challenge for interactive PoRep is drawn used to ensure it is not predictable by miner.

PRE_COMMIT_SECTOR_BATCH_MAX_SIZE

The maximum number of sector pre-commitments in a single batch. 32 sectors per epoch would support a single miner onboarding 1 EiB of 32 GiB sectors in 1 year.

PROVE_REPLICA_UPDATES_MAX_SIZE

The maximum number of sector replica updates in a single batch. Same as PRE_COMMIT_SECTOR_BATCH_MAX_SIZE for consistency

PROV_COLLATERAL_PERCENT_SUPPLY_DENOM

Denominator of the percentage of normalized circulating supply that must be covered by provider collateral

PROV_COLLATERAL_PERCENT_SUPPLY_NUM

Numerator of the percentage of normalized circulating supply that must be covered by provider collateral

SECTORS_MAX

The maximum number of sectors that a miner can have simultaneously active. This also bounds the number of faults that can be declared, etc.

WORKER_KEY_CHANGE_DELAY

Staging period for a miner worker key change. Finality is a harsh delay for a miner who has lost their worker key, as the miner will miss Window PoSts until it can be changed. It’s the only safe value, though. We may implement a mitigation mechanism such as a second key or allowing the owner account to submit PoSts while a key change is pending.

WPOST_CHALLENGE_LOOKBACK

Lookback from the deadline’s challenge window opening from which to sample chain randomness for the challenge seed. This lookback exists so that deadline windows can be non-overlapping (which make the programming simpler) but without making the miner wait for chain stability before being able to start on PoSt computation. The challenge is available this many epochs before the window is actually open to receiving a PoSt.

WPOST_CHALLENGE_WINDOW

The duration of a deadline’s challenge window, the period before a deadline when the challenge is available.

WPOST_DISPUTE_WINDOW

WPOST_MAX_CHAIN_COMMIT_AGE

The maximum distance back that a valid Window PoSt must commit to the current chain.

WPOST_PERIOD_DEADLINES

The number of non-overlapping PoSt deadlines in each proving period.

WPOST_PROVING_PERIOD

The period over which all a miner’s active sectors will be challenged.