Module fil_actors_shared::v13::util

Modules

cbor

Structs

BatchReturn

BatchReturnGen

EventBuilder

Builder for ActorEvent objects, accumulating key/value pairs.

FailCode

Map2

Wraps a HAMT to provide a convenient map API. Any errors are returned with exit code indicating illegal state. The name is not persisted in state, but adorns any error messages.

MapMap

MessageAccumulator

Accumulates a sequence of messages (e.g. validation failures).

Multimap

Multimap stores multiple values per key in a Hamt of Amts. The order of insertion of values for each key is retained.

Set

Set is a Hamt with empty values for the purpose of acting as a hash set.

SetMultimap

SetMultimap is a hamt with values that are also a hamt but are of the set variant. This allows hash sets to be indexable by an address.

Constants

DEFAULT_HAMT_CONFIG

Traits

ActorDowncast

Trait to allow multiple error types to be able to be downcasted into an ActorError.

MapKey

Functions

stack

Computes a batch return that is the result of a sequence of batch returns applied to the previous successful results. Each batch’s size() must be equal to the previous batch’s success_count. Any fail codes then override the prior stack’s successful items, indexed against only those successful items. E.g. stack([OK, E1, OK, E2], [OK, E3], [E4]) => [E4, E1, E3, E2]

Type Aliases

Config