Trait fil_actors_shared::v13::vm_api::VM

pub trait VM {

    // Required methods
    fn blockstore(&self) -> &dyn Blockstore;
    fn actor(&self, address: &Address) -> Option<ActorState>;
    fn set_actor(&self, key: &Address, a: ActorState);
    fn balance(&self, address: &Address) -> TokenAmount;
    fn resolve_id_address(&self, address: &Address) -> Option<Address>;
    fn execute_message(
        &self,
        from: &Address,
        to: &Address,
        value: &TokenAmount,
        method: MethodNum,
        params: Option<IpldBlock>
    ) -> Result<MessageResult, VMError>;
    fn execute_message_implicit(
        &self,
        from: &Address,
        to: &Address,
        value: &TokenAmount,
        method: MethodNum,
        params: Option<IpldBlock>
    ) -> Result<MessageResult, VMError>;
    fn take_invocations(&self) -> Vec<InvocationTrace>;
    fn primitives(&self) -> &dyn Primitives;
    fn mut_primitives(&self) -> &dyn MockPrimitives;
    fn actor_manifest(&self) -> BTreeMap<Cid, Type>;
    fn actor_states(&self) -> BTreeMap<Address, ActorState>;
    fn epoch(&self) -> ChainEpoch;
    fn set_epoch(&self, epoch: ChainEpoch);
    fn circulating_supply(&self) -> TokenAmount;
    fn set_circulating_supply(&self, supply: TokenAmount);
    fn base_fee(&self) -> TokenAmount;
    fn set_base_fee(&self, amount: TokenAmount);
    fn timestamp(&self) -> u64;
    fn set_timestamp(&self, timestamp: u64);

}

An abstract VM that is injected into integration tests

Required Methods

fn blockstore(&self) -> &dyn Blockstore

Returns the underlying blockstore of the VM

fn actor(&self, address: &Address) -> Option<ActorState>

Get information about an actor

fn set_actor(&self, key: &Address, a: ActorState)

Upsert an actor into the state tree

fn balance(&self, address: &Address) -> TokenAmount

Get the balance of the specified actor

fn resolve_id_address(&self, address: &Address) -> Option<Address>

Get the ID for the specified address

fn execute_message( &self, from: &Address, to: &Address, value: &TokenAmount, method: MethodNum, params: Option<IpldBlock> ) -> Result<MessageResult, VMError>

Send a message between the two specified actors

fn execute_message_implicit( &self, from: &Address, to: &Address, value: &TokenAmount, method: MethodNum, params: Option<IpldBlock> ) -> Result<MessageResult, VMError>

Send a message without charging gas

fn take_invocations(&self) -> Vec<InvocationTrace>

Take all the invocations that have been made since the last call to this method

fn primitives(&self) -> &dyn Primitives

Provides access to VM primitives

fn mut_primitives(&self) -> &dyn MockPrimitives

Provides access to VM primitives that can be mocked

fn actor_manifest(&self) -> BTreeMap<Cid, Type>

Return a map of actor code CIDs to their corresponding types

fn actor_states(&self) -> BTreeMap<Address, ActorState>

Returns a map of all actor addresses to their corresponding states

fn epoch(&self) -> ChainEpoch

Get the current chain epoch

fn set_epoch(&self, epoch: ChainEpoch)

Sets the epoch to the specified value

fn circulating_supply(&self) -> TokenAmount

Get the circulating supply constant for the network

fn set_circulating_supply(&self, supply: TokenAmount)

Set the circulating supply constant for the network

fn base_fee(&self) -> TokenAmount

Get the current base fee

fn set_base_fee(&self, amount: TokenAmount)

Set the current base fee

fn timestamp(&self) -> u64

Get the current timestamp

fn set_timestamp(&self, timestamp: u64)

Set the current timestamp