Module fil_actors_shared::v14::builtin::shared

Structs

SendError

An error returned on a failed send. Can be automatically converted into an ActorError with the question-mark operator.

Constants

FIRST_ACTOR_SPECIFIC_EXIT_CODE

FIRST_EXPORTED_METHOD_NUMBER

HAMT_BIT_WIDTH

Functions

extract_send_result

resolve_to_actor_id

ResolveToActorID resolves the given address to its actor ID. If an actor ID for the given address doesn’t exist yet, it tries to create one by sending a zero balance to the given address.

restrict_internal_api