Trait fil_actors_shared::v14::runtime::MessageInfo

pub trait MessageInfo {
    // Required methods
    fn nonce(&self) -> u64;
    fn caller(&self) -> Address;
    fn origin(&self) -> Address;
    fn receiver(&self) -> Address;
    fn value_received(&self) -> TokenAmount;
    fn gas_premium(&self) -> TokenAmount;
}

Message information available to the actor about executing message.

Required Methods

fn nonce(&self) -> u64

The nonce of the currently executing message.

fn caller(&self) -> Address

The address of the immediate calling actor. Always an ID-address.

fn origin(&self) -> Address

The address of the origin of the current invocation. Always an ID-address

fn receiver(&self) -> Address

The address of the actor receiving the message. Always an ID-address.

fn value_received(&self) -> TokenAmount

The value attached to the message being processed, implicitly added to current_balance() before method invocation.

fn gas_premium(&self) -> TokenAmount

The message gas premium