Module fil_actors_shared::v9::util

Structs

BatchReturn

FailCode

MapMap

Multimap

Multimap stores multiple values per key in a HAMT of Amts. The order of insertion of values for each key is retained.

Set

Set is a HAMT with empty values for the purpose of acting as a hash set.

SetMultimap

SetMultimap is a HAMT with values that are also a HAMT but are of the set variant. This allows hash sets to be indexable by an address.

Traits

ActorDowncast

Trait to allow multiple error types to be able to be downcasted into an ActorError.