Trait fil_actors_shared::v9::util::ActorDowncast

pub trait ActorDowncast {
    // Required methods
    fn downcast_default(
        self,
        default_exit_code: ExitCode,
        msg: impl AsRef<str>
    ) -> ActorError;
    fn downcast_wrap(self, msg: impl AsRef<str>) -> Error;
}

Trait to allow multiple error types to be able to be downcasted into an ActorError.

Required Methods

fn downcast_default( self, default_exit_code: ExitCode, msg: impl AsRef<str> ) -> ActorError

Downcast a dynamic std Error into an ActorError. If the error cannot be downcasted into an ActorError automatically, use the provided ExitCode to generate a new error.

fn downcast_wrap(self, msg: impl AsRef<str>) -> Error

Wrap the error with a message, without overwriting an exit code.

Object Safety

This trait is not object safe.